More code clean up.
