+++ og.info.yml
@@ -9,4 +9,3 @@ dependencies:
   - options
   - text
   - user
-  - views
