+++ og_ui/og_ui.info.yml
@@ -7,5 +7,6 @@ core: 8.x
 
 dependencies:
   - og
+  - views
 
 configure: og_ui.admin_index
